Change rendering to use ushort indices, not uint
In OpenGL ES, glDrawElements should use unsigned short to be in spec,
not unsigned ints.  Thus, the indices have to be changed back to be
shorts.

+++ src/material_manager.cpp
@@ -173,7 +173,7 @@ Mesh *MaterialManager::LoadMesh(const char *filename) {
       auto mat = LoadMaterial(surface->material()->c_str());
       if (!mat) { delete mesh; return nullptr; }  // Error msg already set.
       mesh->AddIndices(
-          reinterpret_cast<const uint32_t *>(surface->indices()->Data()),
+          reinterpret_cast<const uint16_t *>(surface->indices()->Data()),
           surface->indices()->Length(), mat);
     }
     mesh_map_[filename] = mesh;
